refactor(stylelint-plugin-meteor): move meteor/prefer-color-token rule into rules directory
